requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
django-rest-multiple-models==2.1.3
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.2
